+++ src/constants.js
@@ -6,6 +6,7 @@ export const DUNGEON_VERSION = "1.0";
 
 export const FLAG_DUNGEON_VERSION = "dungeonVersion";
 
+export const SETTING_ALLOW_TRUSTED_PLAYER = "allowTrustedPlayer";
 export const SETTING_3DCANVAS_ENABLED = "3DCanvasEnabled";
 export const SETTING_CUSTOM_THEMES = "customThemes";
 export const SETTING_RELEASE_NOTES_VERSION = "releaseNotesVersion";